backend/requirements.txt
fastapi[all]
uvicorn[standard]

sqlalchemy[asyncio]
asyncpg
alembic
psycopg[binary]

redis
httpx

pydantic
pydantic-settings
python-dotenv

openai
langchain
langchain-openai

python-jose[cryptography]
cryptography

structlog
slowapi

pytest
pytest-asyncio
pytest-cov
mypy
ruff
